@@ -23,18 +23,19 @@
   "dependencies": {
     "@ajsf/bootstrap4": "^0.1.4",
     "@ajsf/material": "^0.1.4",
-    "@angular/animations": "~8.2.14",
-    "@angular/cdk": "~8.2.3",
-    "@angular/common": "~8.2.14",
-    "@angular/compiler": "~8.2.14",
-    "@angular/core": "~8.2.14",
-    "@angular/flex-layout": "^8.0.0-beta.27",
-    "@angular/forms": "~8.2.14",
-    "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "~8.2.14",
-    "@angular/platform-browser-dynamic": "~8.2.14",
-    "@angular/router": "~8.2.14",
-    "@tibco/spotfire-wrapper": "^0.7.3",
+    "@angular/animations": "^9.1.12",
+    "@angular/cdk": "^9.2.4",
+    "@angular/common": "^9.1.12",
+    "@angular/compiler": "^9.1.12",
+    "@angular/core": "^9.1.12",
+    "@angular/flex-layout": "^9.0.0-beta.31",
+    "@angular/forms": "^9.1.12",
+    "@angular/material": "^9.2.4",
+    "@angular/platform-browser": "^9.1.12",
+    "@angular/platform-browser-dynamic": "^9.1.12",
+    "@angular/router": "^9.1.12",
+    "@ngtools/webpack": "^9.1.11",
+    "@tibco/spotfire-wrapper": "^0.8.0",
     "brace": "^0.11.1",
     "chart.js": "^2.9.3",
     "chartjs-plugin-datalabels": "^0.7.0",
@@ -43,30 +44,30 @@
     "document-register-element": "^1.14.3",
     "hammerjs": "^2.0.8",
     "lodash-es": "^4.17.15",
-    "ng2-charts": "^2.3.0",
+    "ng2-charts": "^2.3.2",
     "ngx-color-picker": "^8.2.0",
-    "npm": "^6.14.5",
-    "primeicons": "^4.0.0-rc.2",
-    "primeng": "^9.1.0",
-    "rxjs": "^6.5.5",
+    "npm": "^6.14.6",
+    "primeicons": "^4.0.0",
+    "primeng": "^9.1.3",
+    "rxjs": "^6.6.0",
     "tslib": "^1.11.2",
-    "zone.js": "~0.9.1"
+    "zone.js": "^0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.26",
-    "@angular-devkit/build-ng-packagr": "^0.803.26",
-    "@angular/cli": "^8.3.26",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/elements": "^8.2.14",
-    "@angular/language-service": "~8.2.14",
-    "@types/node": "^12.12.38",
+    "@angular-devkit/build-angular": "^0.901.11",
+    "@angular-devkit/build-ng-packagr": "^0.901.11",
+    "@angular/cli": "^9.1.11",
+    "@angular/compiler-cli": "^9.1.12",
+    "@angular/elements": "^9.1.12",
+    "@angular/language-service": "^9.1.12",
+    "@types/node": "^12.12.50",
     "codelyzer": "^5.2.2",
     "concat": "^1.0.3",
     "fs-extra": "^8.1.0",
-    "ng-packagr": "^5.7.1",
+    "ng-packagr": "^9.1.5",
     "ts-node": "~8.5.4",
     "tsickle": "^0.37.1",
     "tslint": "~5.20.1",
-    "typescript": "~3.4.5"
+    "typescript": "^3.8.3"
   }
 }
