@@ -5,12 +5,12 @@
     "compodoc": "npx compodoc -p tsconfig.lib.json"
   },
   "peerDependencies": {
-    "@angular/common": "^8.2.14",
-    "@angular/core": "^8.2.14",
-    "@angular/material": "^8.2.3",
-    "@angular/flex-layout": "^8.0.0-beta.27",
+    "@angular/common": "^9.1.12",
+    "@angular/core": "^9.1.12",
+    "@angular/material": "^9.2.4",
+    "@angular/flex-layout": "^9.0.0-beta.31",
     "ngx-color-picker": "^8.2.0",
-    "ng2-charts": "^2.3.0",
+    "ng2-charts": "^2.3.2",
     "chartjs-plugin-datalabels": "^0.7.0",
     "chartjs-plugin-doughnutlabel": "^2.0.3",
     "@tibco-tcstk/tc-core-lib": "^1.1.1",
