@@ -1,20 +1,16 @@
 import {
-  AfterViewChecked,
   AfterViewInit,
   Component,
   ElementRef,
   EventEmitter,
   Input,
-  OnChanges,
-  OnDestroy,
   OnInit,
   Output,
   ViewChild
 } from '@angular/core';
 import {LiveAppsService} from '../../services/live-apps.service';
-import {Subject} from 'rxjs';
-import {map, take, takeUntil} from 'rxjs/operators';
-import {CaseList, CaseRoute} from '../../models/liveappsdata';
+import { take, takeUntil} from 'rxjs/operators';
+import { CaseRoute} from '../../models/liveappsdata';
 import {LiveAppsComponent} from '../live-apps-component/live-apps-component.component';
 import {TcComponent, TcCoreCommonFunctions} from '@tibco-tcstk/tc-core-lib';
 
@@ -42,13 +38,24 @@ export class LiveAppsRecentCasesComponent extends LiveAppsComponent implements O
   @Input() uiAppId: string;
 
   /**
-   * case card format - list, card, miniCard, staticList (no click event)
+   * case card format - list, card, miniCard, staticList (no click event):  miniCard, card, list
    */
-  @Input() displayType: string = this.displayType ? this.displayType : 'miniCard'; // miniCard, card, list
+  public displayType: string = 'miniCard';
+  @Input('displayType') set DisplayType(displayType: string) {
+    if (displayType){
+      this.displayType = displayType;
+    }
+  }
+
   /**
    * Whether to show the header bar in the widget - eg. favorites on home page (contains icon etc) - if off icons still appear without bar
    */
-  @Input() showHeader: boolean = this.showHeader ? this.showHeader : true;
+  public showHeader: boolean = true;
+  @Input('showHeader') set ShowHeader(showHeader: boolean) {
+    if (showHeader){
+      this.showHeader = showHeader;
+    }
+  }
 
   /**
    * ~event clickCase : Case clicked
