@@ -1,7 +1,8 @@
-import {Component, Input, OnChanges, OnInit, SimpleChanges, ViewChild} from '@angular/core';
-import {MatPaginator, MatSort} from '@angular/material';
-import {TibcoCloudTableDataSource} from './tibco-cloud-table-datasource';
-import {TcCoreCommonFunctions} from '../../common/tc-core-common-functions';
+import { Component, Input, OnChanges, OnInit, ViewChild } from '@angular/core';
+import { MatPaginator } from '@angular/material/paginator';
+import { MatSort } from '@angular/material/sort';
+import { TibcoCloudTableDataSource } from './tibco-cloud-table-datasource';
+import { TcCoreCommonFunctions } from '../../common/tc-core-common-functions';
 
 @Component({
   selector: 'tc-tibco-cloud-table',
