@@ -2,15 +2,18 @@ import { NgModule, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { MaterialDesignFrameworkModule } from '@ajsf/material';
 import { Bootstrap4FrameworkModule } from '@ajsf/bootstrap4';
 import {RenderedFormComponent} from './components/rendered-form/rendered-form.component';
-import {
-  MatButtonModule,
-  MatCardModule,
-  MatCheckboxModule, MatDialogModule,
-  MatFormFieldModule,
-  MatIconModule,
-  MatInputModule,
-  MatListModule, MatMenuModule, MatOptionModule, MatSelectModule, MatToolbarModule, MatTooltipModule
-} from '@angular/material';
+import {MatButtonModule} from '@angular/material/button';
+import {MatCardModule} from '@angular/material/card';
+import {MatCheckboxModule} from '@angular/material/checkbox';
+import {MatDialogModule} from '@angular/material/dialog';
+import {MatFormFieldModule} from '@angular/material/form-field';
+import {MatIconModule} from '@angular/material/icon';
+import {MatInputModule} from '@angular/material/input';
+import {MatListModule} from '@angular/material/list';
+import {MatMenuModule} from '@angular/material/menu';
+import {MatSelectModule} from '@angular/material/select';
+import {MatToolbarModule} from '@angular/material/toolbar';
+import {MatTooltipModule} from '@angular/material/tooltip';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import {FlexLayoutModule} from '@angular/flex-layout';
 import {ScrollingModule} from '@angular/cdk/scrolling';
@@ -43,7 +46,7 @@ import {CommonModule} from '@angular/common';
     MatInputModule,
     MatIconModule,
     MatSelectModule,
-    MatOptionModule,
+    // MatOptionModule,
     MatDialogModule,
     MatMenuModule,
     MatCardModule,
