@@ -1,7 +1,7 @@
 import {Component, EventEmitter, Input, OnChanges, OnInit, Output, SimpleChanges, ViewChild} from '@angular/core';
 import {CaseType, JsonSchema, Process} from '../../models/liveappsdata';
 import {TcCaseProcessesService} from '../../services/tc-case-processes.service';
-import {MatExpansionPanel} from '@angular/material';
+import {MatExpansionPanel} from '@angular/material/expansion';
 import {FormConfig, ProcessFormConfig} from '../../models/tc-liveapps-config';
 import {TcCoreCommonFunctions} from '@tibco-tcstk/tc-core-lib';
 import {RenderedFormComponent} from '@tibco-tcstk/tc-forms-lib';
@@ -20,7 +20,12 @@ export class LiveAppsFormPreviewComponent implements OnChanges {
   @Input() appIds: string[];
   @Input() sandbox: string;
   @Input() formConfig: FormConfig;
-  @Input() formsFramework = this.formsFramework ? this.formsFramework : 'material-design';
+  public formsFramework = 'material-design';
+  @Input('formsFramework') set FormsFramework(formsFramework) {
+    if (formsFramework){
+      this.formsFramework = formsFramework;
+    }
+  }
 
   /**
    * ~event saveFormConfiguration : save form config
