@@ -52,13 +52,22 @@ export class LiveAppsWorkitemsComponent extends LiveAppsComponent implements OnI
   /**
    * case card format - list, card, miniCard, staticList (no click event)
    */
-  @Input() displayType: string = this.displayType ? this.displayType : 'wiMiniCard'; // miniCard, card, list
+  public displayType: string = 'wiMiniCard';
+  @Input('displayType') set DisplayType(displayType: string) {
+    if (displayType){
+      this.displayType = displayType;
+    }
+  }
+
   /**
    * Whether to show the header bar in the widget - eg. favorites on home page (contains icon etc) - if off icons still appear without bar
    */
-  @Input() showHeader: boolean = this.showHeader ? this.showHeader : true;
-
-
+  public showHeader: boolean = true;
+  @Input('showHeader') set ShowHeader(showHeader: boolean) {
+    if (showHeader){
+      this.showHeader = showHeader;
+    }
+  }
 
   /**
    * ~event clickCase : Case clicked
