@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { MatSnackBar } from '@angular/material';
+import { MatSnackBar } from '@angular/material/snack-bar';
 import { TcSpotfireConfigService } from '../../services/tc-spotfire-config.service';
 import { SpotfireConfig } from '../../models/tc-spotfire-config';
 
