@@ -7,15 +7,24 @@ import {Location, CurrencyPipe, DatePipe} from '@angular/common';
 import { RouterModule } from '@angular/router';
 import {FlexLayoutModule, FlexModule} from '@angular/flex-layout';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
-import {
-  MatButtonModule, MatButtonToggleModule,
-  MatCardModule,
-  MatCheckboxModule, MatDialogModule,
-  MatExpansionModule,
-  MatFormFieldModule, MatIconModule, MatInputModule,
-  MatListModule, MatMenuModule, MatOptionModule, MatSelectModule,
-  MatTabsModule, MatToolbarModule, MatTooltipModule
-} from '@angular/material';
+
+import {MatButtonModule} from '@angular/material/button';
+import {MatButtonToggleModule} from '@angular/material/button-toggle';
+import {MatCardModule} from '@angular/material/card';
+import {MatCheckboxModule} from '@angular/material/checkbox';
+import {MatDialogModule} from '@angular/material/dialog';
+import {MatExpansionModule} from '@angular/material/expansion';
+import {MatFormFieldModule} from '@angular/material/form-field';
+import {MatIconModule} from '@angular/material/icon';
+import {MatInputModule} from '@angular/material/input';
+import {MatListModule} from '@angular/material/list';
+import {MatMenuModule} from '@angular/material/menu';
+import {MatSelectModule} from '@angular/material/select';
+import {MatTabsModule} from '@angular/material/tabs';
+import {MatToolbarModule} from '@angular/material/toolbar';
+import {MatTooltipModule} from '@angular/material/tooltip';
+
+
 import {
   CachingInterceptor,
   LogService,
@@ -94,7 +103,7 @@ const tcCoreConfig: TcCoreConfig = {
     MatToolbarModule,
     MatIconModule,
     MatSelectModule,
-    MatOptionModule,
+    // MatOptionModule,
     MatDialogModule,
     MatMenuModule,
     MatCardModule,
